Significant Refactoring: Signals like SegmentEndPointSignal used to carry Data inside it. Now, the Data has been separated from the Signal, so a Signal is purely a Signal. Consequently, the Data/Signal stream would now looks like:
<SegmentEndPointSignal> <AudioFrame> ... <AudioFrame> <SegmentEndPointSignal>

instead of

<SegmentEndPointSignal> <AudioFrame>...<AudioFrame> <SegmentEndPointSignal>
   |-> <AudioFrame>                                    |-> <AudioFrame>

Not only is this approach conceptually better, but also (contrary to my
initial believe) makes the implementation of the processors a lot cleaner.
There is a lot less type casting and requires no indirection
to get to actual Data object inside the Signal. I'm happy with this
refactoring :) Although it requires some look ahead in the FeatureExtractor
(which is handled by a peekable InputQueue), I believe this is the right
approached.

All previous tests passed.


git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@323 c60273ff-4a12-0410-8d58-82ceea6d4170