Added program & script to do the Sphinx 3 metrics.
To run the script, read tests/isolatedDigits-ti46.sh. For SUN people,
the script should just run fine. For people outside SUN, you need
to change the location of the isolated digits model data.

In libs3decoder/utt.c, the output is written to stderr instead of stdout
in the function utt_end().


git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk/sphinx3@458 c60273ff-4a12-0410-8d58-82ceea6d4170