Added convience static versions of 'start' and 'stop' that take a name of a timer. The timer will be automatically created and used. Timer.dumpAll will dump these timers.
This eliminates the need to have to create timers when you want to do
a quick timing task.  Before you would have to to something like this:


Timer myTimer = Timer(context, "loop");

myTimer.start()
runLoop();
myTimer.stop()

Timer.dumpAll();


Now it is possible to do this:

myTimer.start("loop")
runLoop();
myTimer.stop("loop")

Timer.dumpAll();

this is more convienient when there are several timers that need to be
managed.


git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@1082 c60273ff-4a12-0410-8d58-82ceea6d4170