Added to the FrontEnd interface the method setDataSource(DataSource). Implemented this method in BaseFrontEnd.java and ParallelFrontEnd.java. This allows switching the data to decode on the fly.
BaseFrontEnd.java uses StubDataSource objects, which encapsulates the
real DataSource objects. The StubDataSource is now the DataSource for the
first processor in the pipeline. This way, changing the real DataSource only
requires calling StubDataSource.setDataSource(). It does not require
changing the DataSource for the first processor in the pipeline. This
avoids adding setDataSource() methods to all processors.

The ParallelFrontEnd.setDataSource() works in almost the same way.


git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk/sphinx4@1460 c60273ff-4a12-0410-8d58-82ceea6d4170