Modified the ThreadedAcousticScorer to reduce the threading overhead. This gave an overall 7% increase in speed on the ripley test set. With this test set, single CPU scoring takes 160s. With the previous threaded scorer using 2 CPUs, scoring took 135s, which implies that CPUs are not used to their full potential. With this modified scorer, scoring took 110s (+18%).
The original scorer works as follows. If you have 2 CPUs, 2 scoring threads
will be created, in addition to the main calling thread (3 threads in total).
However, the main calling thread just waits for the scoring threads to finish,
and does not do any scoring.

The modified scorer, given 2 CPUs, creates only 1 additional thread (2 threads
in total). The original calling thread is used for scoring as well,
therefore reducing the overhead of having an extra thread.

This modified scorer gives the an4 and rm1 tests a similar performance
improvement on mangueria.


git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@1888 c60273ff-4a12-0410-8d58-82ceea6d4170