Increased the 'startOffset' property default from 5 frames to 10 frames, same as the 'endOffset'.
The 'startOffset' is the number of frames before the start of speech that
is included as part of the speech (i.e., the pre-speech silence). 5 frames is
just ridiculously short. Tested this out on the live demo, and the missing
initial 'the's are not happening any more :) The property is:

edu.cmu.sphinx.frontend.EnergyEndpointer.startOffset


git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@2162 c60273ff-4a12-0410-8d58-82ceea6d4170