Changed the search string for grep to "FWDVIT".
Changed the test-cepview target so that it checks if the diff was
successful (i.e. if the files are equal) and prints out a simple
message "passed" or "failed" instead of the whole diff in the case
when the files differ. Volunteers to do the same with other targets
are most welcome.

Changed the CLEANFILES list to use a wildcard. Keep in mind that if you
create a new target and the target produces files, you have to add the
filenames to this CLEANFILES list, so they're removed when one types
'make clean'


git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@4165 c60273ff-4a12-0410-8d58-82ceea6d4170