Replaced "clean:" target with CLEANFILES variable (since we're using automake, the clean target overwrites the default)
Removed "-cepdir" where not needed (e.g. "-cepdir /" when the control
file already has the "/")

Replaced the alignment program with sclite. Beware that sclite has to
be on your path, or on the path of the script running the regression
test. Hardwiring a path would cause the code to break in a different
platform than linux. A default path is ~robust/archive/third_party_packages/NIST_scoring_tools/sctk/linux/bin

Added "-rawext" to some of the livepretend tests, so that the
reference and hypothesis files would have exactly the same file names
(last item in each line, surrounded by ())


git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk/sphinx3@4356 c60273ff-4a12-0410-8d58-82ceea6d4170