Merged from branch SPHINX3_5_2_RCI_IRII_BRANCH:
1, Added Letter-to-sound LTS rule, dict_init will only specify
d->lts_rules to be true if the useLTS is specified.  Only if
d->lts_rules is specified, the LTS logic will be used. The code safe
guarded the case when a phone in mdef doesn't appear in LTS, in that
case, the code will force exit.

2, The LTS logic is only used as a reserved measure.  By default, it
is not turned on.  See also the comment in kbcore.c and the default
parameters in revision 1.3 cmdln_macro.h . We added it because we have
this functionality in SphinxTrain.


git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk/sphinx3@5283 c60273ff-4a12-0410-8d58-82ceea6d4170