add -outputfullpath and -fullsuffixmatch arguments to bw.
Default behavior is to keep the existing system behavior when the
corpus module tries to match the transcript utterance id with the
partial path contained in the control file.

Using -fullsuffixmatch yes will do the following:
	The corpus module will check whether the string contained
	inside parentheses in the transcript for the utterances
	matches the final part of the control file partial path
	for the utterance.  For instance, if the control file
	partial path is:
		tidigits/train/man/ae/243za
	the following strings will be considered to match:
		243za
		ae/243za
		man/ae/243za
		.
		.
		.
	In any event, the utterance will be used by bw for training.
	This switch just modifies when the warning message for
	mismatching control file and transcripts is generated.

Using -outputfullpath yes will output the entire subpath from the
control file in the log output of bw rather than just the final path
component.  This allows for simpler automatic processing of the output
of bw.


git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@5334 c60273ff-4a12-0410-8d58-82ceea6d4170