Clean up the log handling. In logs3.c, removed unnecessary variables (e.g. "f", exactly the same as "F") and functions (e.g. "logs3_10base()").
In confidence.c, replace (logs3_to_log10(r_lscr) * logs3_10base())
with r_lscr, since the only difference is that one is a double, the
other an int (and as such, they differ on the order of 1e-12).

In future cleanups.... replace the "int" declaration with "int32",
used in the rest of the code.


git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@5446 c60273ff-4a12-0410-8d58-82ceea6d4170