Summary: Fixed another 32bit LM issue.  LM Conversion is fine. Decoding is yet to be test. Also added another 7 regression tests.
Details: There exists some LM with unigram smaller than 0xffff in
which when the segment size reach BG_SEG_SZ, the tg_count could more
than 0xffff. That simply means on avg, there are more than
(0xffff/BG_SEG_SZ) words follows each bigram in one segment.  

Example, if BG_SEG_SZ is 512, if on average there is more than 128
words follow these 512 bigrams, by the design of the old LM.  This
will break the limit of the old code.

Current treatment is if the lm reading routine see this situation, it
will backoff the 32bit LM routine where there is no such limitation at
all (first_tg could be larger than 0xffff).  

We don't know the effect of this test in decoding yet.  Will test it
when I wake up tmw. 

Also introduced a mode "TXT32" that could force lm_convert to convert
a txt LM and treat it like a 32 bit LM.  The 7 regression tests were
added for that reason.

A note on the rationale of changing the LM code: inside liblm, they
are now fairly messed up.  But I do insist on make the interface
clean. The only exception is flat_fwd.c.  My plan is to use macros to
clean up the multiple data structures issue later on.  This code is ok
as long as we continuously test the code using more difficult test
cases.




git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@5814 c60273ff-4a12-0410-8d58-82ceea6d4170