A bunch of minor changes to the front-end.  The old-fashioned MFCC is not affected and continues to do as it always did albeit with a tiny bit more accuracy and speed on fixed-point platforms.  The question of exactly which DCT is being used for what has been resolved with some comments to that effect in the header.
The misnamed and misfunctional -idct and -dct options to sphinx_fe
have been removed and replaced with -dct2 and -dct3, which do the
canonical (and invertible) DCT-II and DCT-III.  An option -smoothspec
has been added to output the cepstral-smoothed log spectrum.

Some tests have been added for these new features.


git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk/sphinxbase@6115 c60273ff-4a12-0410-8d58-82ceea6d4170