Some tyding of HMM evaluation.  The big change here is that multiplex triphones now use -1 to indicate that a given state hasn't been reached yet, and therefore we don't repeatedly evaluate senone sequence #0 (which could be any arbitrary thing) vacuously.
This also allows us to make the assumption that all multiplexed
senones share the same transition matrix (which is the case by
definition since they all share the same base phone), which reduces
the number of variables and (hopefully) leads to better object code.

Also use real from and to indices instead of cryptic arc numbers in
NPA().

This is an intermediate step towards eliminating the SMD structure and
using the transition matrices directly.


git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@6237 c60273ff-4a12-0410-8d58-82ceea6d4170