Finally, add an option to remove any DC offset at the frame level, equivalent to the ZMEANSOURCE option to HCopy.
This allows us to come within a couple percent of HTK features (the
difference probably results from numerical errors), *except* for the
power coefficient, because HTK's DCT-II does not apply the
orthogonalizing factor of 1/sqrt(2) to it.  If it's worth being
exactly compatible I guess we can add a transform type "htk".

Any differences with HTK at this point probably result from the way in
which CMN and dynamic features are done.


git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@6841 c60273ff-4a12-0410-8d58-82ceea6d4170