Implement lattice generation from FSG search.  The bit about finding the start and end nodes in the lattice is not quite correct yet, I need to figure out a proof of the really correct way to do this.
There is also the matter of the implicit link exiting the final node.

N-best and posterior remain to be implemented but they are not
complicated.  Of course, the lattices generated by the unit tests are
comically sparse, so we will need a more interesting test in order to
verify that they actually give meaningful results :)


git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@8261 c60273ff-4a12-0410-8d58-82ceea6d4170