Added an example Groovy script.  The script is located in a new directory src/scripts/location
Here is the README

This directory contains a example Groovy script.  This example reimplements the Transciber application. 

Groovy is a popular choice for Java based scripting, and is a good choice for rapid development with Sphinx4.

To try this script, install Groovy from here

    http://groovy.codehaus.org/

Then cd to src/scripts/groovy and type the following to the shell

    groovy -cp ../../../lib/sphinx4.jar:../../../lib/jsapi.jar GroovyTranscriber.groovy ../../../




git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@9416 c60273ff-4a12-0410-8d58-82ceea6d4170