Adding a Clojure script example.  Here is the README
This directory contains a example Clojure script.  This example reimplements the Transciber application.

Clojure is a popular choice for Java based scripting, and is a good choice for rapid development with Sphinx4.

To try this script, install Clojure from here

    http://clojure.org/

Then cd to src/scripts/clojure and type the following to the shell

    java -cp <clojure home>clojure.jar:../../../lib/sphinx4.jar:../../../lib/jsapi.jar clojure.lang.Repl ClojureTranscriber.clj




git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@9473 c60273ff-4a12-0410-8d58-82ceea6d4170